+++ contracts/staking/src/contract.rs
@@ -412,23 +412,22 @@ mod tests {
     use std::str::FromStr;
 
     fn sample_validator(addr: &str) -> Validator {
-        Validator {
-            address: addr.to_owned(),
-            commission: Decimal::percent(3),
-            max_commission: Decimal::percent(10),
-            max_change_rate: Decimal::percent(1),
-        }
+        Validator::create(
+            addr.to_owned(),
+            Decimal::percent(3),
+            Decimal::percent(10),
+            Decimal::percent(1),
+        )
     }
 
     fn sample_delegation(validator_addr: &str, amount: Coin) -> FullDelegation {
-        let can_redelegate = amount.clone();
-        FullDelegation {
-            validator: validator_addr.to_owned(),
-            delegator: Addr::unchecked(MOCK_CONTRACT_ADDR),
+        FullDelegation::create(
+            Addr::unchecked(MOCK_CONTRACT_ADDR),
+            validator_addr.to_owned(),
+            amount.clone(),
             amount,
-            can_redelegate,
-            accumulated_rewards: Vec::new(),
-        }
+            vec![],
+        )
     }
 
     fn set_validator(querier: &mut MockQuerier) {