+++ packages/std/src/testing/mock.rs
@@ -2216,7 +2216,7 @@ mod tests {
                     if addr == constract1 {
                         let response = ContractInfoResponse {
                             code_id: 4,
-                            creator: "lalala".into(),
+                            creator: Addr::unchecked("lalala"),
                             admin: None,
                             pinned: false,
                             ibc_port: None,
@@ -2235,7 +2235,7 @@ mod tests {
                     if code_id == 4 {
                         let response = CodeInfoResponse {
                             code_id,
-                            creator: "lalala".into(),
+                            creator: Addr::unchecked("lalala"),
                             checksum: HexBinary::from_hex(
                                 "84cf20810fd429caf58898c3210fcb71759a27becddae08dbde8668ea2f4725d",
                             )