+++ packages/std/src/traits.rs
@@ -567,7 +567,7 @@ mod tests {
         fn mock_resp() -> ContractInfoResponse {
             ContractInfoResponse {
                 code_id: 0,
-                creator: "creator".to_string(),
+                creator: Addr::unchecked("creator"),
                 admin: None,
                 pinned: false,
                 ibc_port: None,
@@ -598,7 +598,7 @@ mod tests {
         fn mock_resp() -> ContractInfoResponse {
             ContractInfoResponse {
                 code_id: 0,
-                creator: "creator".to_string(),
+                creator: Addr::unchecked("creator"),
                 admin: None,
                 pinned: false,
                 ibc_port: None,